feat: fitur presensi mahasiswa

+++ app/Http/Controllers/AttendanceController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Attendance;
+use App\Models\Course;
 use Illuminate\Http\Request;
 use App\Models\Lecturer;
 use App\Models\Schedule;
@@ -14,6 +15,7 @@ use RealRashid\SweetAlert\Facades\Alert;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Student;
 use Illuminate\Support\Facades\Schema;
+use PHPUnit\Framework\Constraint\Count;
 
 class AttendanceController extends Controller
 {
@@ -42,7 +44,14 @@ class AttendanceController extends Controller
 
     public function datatable()
     {
-        $model = Attendance::query();
+        if (Auth::user()->hasRole('Mahasiswa')) {
+            $user_id = Auth::user()->id;
+            $student = Student::where('user_id', $user_id)->first();
+
+            $model = Attendance::where('student_id', $student->id)->orderBy('id', 'desc');
+        } else {
+            $model = Attendance::query();
+        }
 
         return DataTables::of($model)
             ->editColumn('created_at', function ($data) {
@@ -56,8 +65,13 @@ class AttendanceController extends Controller
 
                 $btn = "<div class='btn-group'>";
                 $btn .= "<a href='$url_show' class = 'btn btn-outline-info btn-sm text-nowrap'><i class='fas fa-info mr-2'></i> Detail</a>";
-                $btn .= "<a href='$url_edit' class = 'btn btn-outline-info btn-sm text-nowrap'><i class='fas fa-edit mr-2'></i> Edit</a>";
-                $btn .= "<a href='$url_delete' class = 'btn btn-outline-danger btn-sm text-nowrap' data-confirm-delete='true'><i class='fas fa-trash mr-2'></i> Hapus</a>";
+
+                if (Auth::user()->hasRole('Mahasiswa')) {
+                } else {
+                    $btn .= "<a href='$url_edit' class = 'btn btn-outline-info btn-sm text-nowrap'><i class='fas fa-edit mr-2'></i> Edit</a>";
+                    $btn .= "<a href='$url_delete' class = 'btn btn-outline-danger btn-sm text-nowrap' data-confirm-delete='true'><i class='fas fa-trash mr-2'></i> Hapus</a>";
+                }
+
                 $btn .= "</div>";
 
                 return $btn;
@@ -120,5 +134,135 @@ class AttendanceController extends Controller
         $student = Student::where('user_id', $user_id)->first();
 
         return view('attendances.present', compact('data', 'student'));
+    }
+
+    public function store(Request $request)
+    {
+        try {
+            DB::beginTransaction();
+
+            $today = Carbon::now()->format('Y-m-d');
+            $time_today = Carbon::now()->format('H:i:s');
+            $date = $request->schedule_date;
+
+            $time_start = Carbon::parse($request->schedule_time_start);
+            $get_time_end = Carbon::parse($request->schedule_time_end);
+            $time_end = $get_time_end->addMinutes(15);
+
+            $latitude = $request->latitude;
+            $longitude = $request->longitude;
+
+            $latitude_mi = -2.978490812012596;
+            $longitude_mi  = 104.73134556182045;
+
+            $schedule_id = Crypt::decrypt($request->schedule_id);
+            $schedule_course_id = Crypt::decrypt($request->schedule_course_id);
+
+            $course = Course::find($schedule_course_id);
+
+            $user_id = Auth::user()->id;
+            $student = Student::where('user_id', $user_id)->first();
+
+            if ($request->type == Attendance::STATUS_PRESENT) {
+                if ($date == $today) {
+                    if ($time_today >= $time_start || $time_today <= $time_end) {
+                        $jarak = $this->haversine($latitude, $longitude, $latitude_mi, $longitude_mi);
+                        $jarakThreshold = 5;
+                        if ($jarak <= $jarakThreshold) {
+                            Attendance::create([
+                                'student_id' => $student->id,
+                                'student_name' => $student->name,
+                                'schedule_id' => $schedule_id,
+                                'schedule_course_name' => $request->schedule_course_name,
+                                'schedule_lecturer_name' => $request->schedule_lecturer_name,
+                                'total_hour' => $course->total_hour,
+                                'latitude' => $latitude,
+                                'longitude' => $longitude,
+                                'status' => Attendance::STATUS_NOT_CONFIRMED
+                            ]);
+                        } else {
+                            // If Data Error
+                            DB::rollBack();
+
+                            // Alert & Redirect
+                            Alert::toast('Anda harus berada dalam radius 5 meter dari lokasi absen.', 'error');
+                            return redirect()->back();
+                        }
+                    } else {
+                        // If Data Error
+                        DB::rollBack();
+
+                        // Alert & Redirect
+                        Alert::toast('Tidak Dapat Melakukan Presensi', 'error');
+                        return redirect()->back();
+                    }
+                } else {
+                    // If Data Error
+                    DB::rollBack();
+
+                    // Alert & Redirect
+                    Alert::toast('Tidak Dapat Melakukan Presensi', 'error');
+                    return redirect()->back();
+                }
+            } elseif ($request->type == Attendance::STATUS_CANT_PRESENT) {
+                Attendance::create([
+                    'student_id' => $student->id,
+                    'student_name' => $student->name,
+                    'schedule_id' => $schedule_id,
+                    'schedule_course_name' => $request->schedule_course_name,
+                    'schedule_lecturer_name' => $request->schedule_lecturer_name,
+                    'total_hour' => $course->total_hour,
+                    'status' => Attendance::STATUS_NOT_CONFIRMED
+                ]);
+            } else {
+                // Save File
+                if ($file = $request->file('file')) {
+                    $destinationPath = 'assets/files/';
+                    $fileName = "SURAT" . "_" . $student->nim . date('YmdHis') . "." . $file->getClientOriginalExtension();
+                    $file->move($destinationPath, $fileName);
+                }
+
+                Attendance::create([
+                    'student_id' => $student->id,
+                    'student_name' => $student->name,
+                    'schedule_id' => $schedule_id,
+                    'schedule_course_name' => $request->schedule_course_name,
+                    'schedule_lecturer_name' => $request->schedule_lecturer_name,
+                    'total_hour' => $course->total_hour,
+                    'status' => Attendance::STATUS_NOT_CONFIRMED,
+                    'file' => $fileName,
+                    'reason' => $request->reason
+                ]);
+            }
+
+            // Save Data
+            DB::commit();
+
+            // Alert & Redirect
+            Alert::toast('Data Berhasil Disimpan', 'success');
+            return redirect()->route('attendance.create');
+        } catch (\Exception $e) {
+            // If Data Error
+            DB::rollBack();
+
+            // Alert & Redirect
+            Alert::toast('Data Gagal Disimpan', 'error');
+            return redirect()->back()->withInput()->with('error', 'Data Tidak Berhasil Diperbarui' . $e->getMessage());
+        }
+    }
+
+    private function haversine($lat1, $lon1, $lat2, $lon2)
+    {
+        // Menghitung jarak antara dua titik koordinat menggunakan Haversine formula.
+        // Anda dapat menemukan implementasi rumus Haversine di banyak sumber online.
+        // Berikut adalah contoh implementasi sederhana:
+        $r = 6371; // Radius Bumi dalam km
+        $dLat = deg2rad($lat2 - $lat1);
+        $dLon = deg2rad($lon2 - $lon1);
+        $a = sin($dLat / 2) * sin($dLat / 2) + cos(deg2rad($lat1)) * cos(deg2rad($lat2)) * sin($dLon / 2) * sin($dLon / 2);
+        $c = 2 * atan2(sqrt($a), sqrt(1 - $a));
+        $d = $r * $c * 1000; // Mengubah jarak menjadi meter
+
+        return $d;
     }
 }
